% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tooltipsterR}
\alias{tooltipsterR}
\alias{tooltipsterR-package}
\title{Interactive Tooltip htmlwidget}
\description{
htmlwidget wrapper for the excellent tooltipster library for interactive tooltips
}

